+++ player_page/code/include/Machine.h
@@ -14,21 +14,21 @@ namespace MiuiIsTheBest {
 
     class Machine {
     public:
-        short current_position = 0;//当前所在的许用位置
-        short depth = 0;
+        int current_position = 0;//当前所在的许用位置
+        int depth = 0;
         unsigned int cost[5];//使用不同能源的费用
         std::vector<Position> *positions;//所有的容许位置
-        std::vector<short> *parents;//父节点
+        std::vector<int> *parents;//父节点
         std::vector<Start> *children;//子节点
         std::vector<bool> *connect_types;//同父节点的连接状态
-        std::vector<short> *cycled_windows;//从每个父节点过来所用的回环窗口
+        std::vector<int> *cycled_windows;//从每个父节点过来所用的回环窗口
         V_STATUS status = V_STATUS::UNDISCOVERED;//节点状态，拓扑排序时使用
         BoolType type;//仪器类型
         bool is_core = false;//是否位于核心生产线
-        short NumParents() { return static_cast<short>(parents->size()); }
+        int NumParents() { return static_cast<int>(parents->size()); }
 
         void RegainCycleTimes(std::vector<Window> &windows) {
-            for (short &index: *cycled_windows) {
+            for (int &index: *cycled_windows) {
                 if (index != -1) {
                     windows.at(index).rest_cycle_times++;
                     index = -1;
@@ -36,9 +36,9 @@ namespace MiuiIsTheBest {
             }
         }
 
-        short CurrentWindow() { return positions->at(current_position).window; }
+        int CurrentWindow() { return positions->at(current_position).window; }
 
-        short CurrentRegion() { return positions->at(current_position).region; }
+        int CurrentRegion() { return positions->at(current_position).region; }
 
         bool OutOfPosition() { return current_position >= positions->size(); }//是否超出所在位置
 
@@ -61,30 +61,30 @@ namespace MiuiIsTheBest {
             }
 
             positions = new std::vector<Position>;
-            parents = new std::vector<short>;
+            parents = new std::vector<int>;
             children = new std::vector<Start>;
             connect_types = new std::vector<bool>;
-            cycled_windows = new std::vector<short>;
+            cycled_windows = new std::vector<int>;
         }
 
-        explicit Machine(short machine_type_number)
+        explicit Machine(int machine_type_number)
                 : type(BoolType(machine_type_number)) {
 
             positions = new std::vector<Position>;
-            parents = new std::vector<short>;
+            parents = new std::vector<int>;
             children = new std::vector<Start>;
             connect_types = new std::vector<bool>;
-            cycled_windows = new std::vector<short>;
+            cycled_windows = new std::vector<int>;
             for (unsigned int &i: cost) {
                 i = 0;
             }
         }
         Machine() {
             positions = new std::vector<Position>;
-            parents = new std::vector<short>;
+            parents = new std::vector<int>;
             children = new std::vector<Start>;
             connect_types = new std::vector<bool>;
-            cycled_windows = new std::vector<short>;
+            cycled_windows = new std::vector<int>;
             for (int i = 0; i < 5; ++i) {
                 cost[i] = 0;
             }
